refactor: number to bigdecimal

+++ app/src/main/java/com/coinhub/android/data/models/TopUpModel.kt
@@ -1,13 +1,14 @@
 package com.coinhub.android.data.models
 
 import kotlinx.datetime.LocalDate
+import java.math.BigDecimal
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
 
 data class TopUpModel @OptIn(ExperimentalUuidApi::class) constructor(
     val id: Uuid,
     val provider: TopUpProviderEnum,
-    val amount: Long,
+    val amount: BigDecimal,
     val status: TopUpStatusEnum,
     val createdAt: LocalDate,
 )
